requirements.txt
allianceauth-corp-tools>=2.8.0
django>=3.2
